Fix parsing committer dates
The previous date format that was passed to `Time.parse` was in a
slightly non-standard format, resulting in the committer timezone not
being preserved. In some cases, this would manifest as a negative
relative time.

Switching to an ISO-8601 date format gives `Time.parse` a standard input
format, which preserves the committer timezone.

This bug existed for a while, but was only revealed after switching to
custom logic (instead of `%(committerdate:relative)` from Git) for
computing relative time strings.